Move banned_external_code node utils to new file
