+++ lib/utils/node_utils.dart
@@ -22,23 +22,3 @@ String humanReadableNodeType(AstNode? node) {
 
   return 'Node';
 }
-
-/// Extension to get library of the node
-extension SimpleIdentifierParentSourceExtension on SimpleIdentifier {
-  /// Returns library of the node
-  String? get sourceUrl {
-    final parentSource = staticElement?.librarySource;
-    final parentSourceName = parentSource?.uri.toString();
-    return parentSourceName;
-  }
-}
-
-/// Extension to get library of the node
-extension NamedTypeParentSourceExtension on NamedType {
-  /// Returns library of the node
-  String? get sourceUrl {
-    final parentSource = element?.librarySource;
-    final parentSourceName = parentSource?.uri.toString();
-    return parentSourceName;
-  }
-}
